+++ src/one_neutral/throw_struct______/bench.cpp
@@ -3,12 +3,12 @@
 
 int main() {
   Setup();
-  for (uint64_t i = 0; i < WARMUP_ITERATIONS; ++i) {
+  for (uint32_t i = 0; i < WARMUP_ITERATIONS; ++i) {
     caller();
   }
   NOP_SLED;
   TimeLogger logger;
-  for (uint64_t i = 0; i < ITERATIONS; ++i) {
+  for (uint32_t i = 0; i < ITERATIONS; ++i) {
     caller();
   }
 }
